Adds some options to the build script.

+++ build-qt5.sh
@@ -1,13 +1,57 @@
 #!/bin/bash
 
+N_THREADS=30
+BUILD_TYPE="-release"
+DEVELOPER_BUILD=
+
+function usage() {
+    echo "Usage: $0 [-d] [-D] [-j #NUMBER]"
+    echo "       -d  : builds with debug symbols."
+    echo "       -D  : builds in developer mode (do not install anything)."
+    echo "       -j #: builds with # threads (default is $N_THREADS)."
+}
+
+while getopts "h?dDj:" opt; do
+    case $opt in
+        h|\?)
+            usage
+            exit 0
+            ;;
+        d)
+            echo "[$0] Building with debug symbols."
+            BUILD_TYPE="-debug"
+            ;;
+        D)
+            echo "[$0] Developer build enabled."
+            DEVELOPER_BUILD=1
+            ;;
+        j)
+            echo "[$0] Building with $OPTARG threads."
+            N_THREADS=$OPTARG
+            ;;
+    esac
+done
+
 RELDIR=`dirname $0`
 ABSDIR=`cd $RELDIR;pwd`
 
 . $ABSDIR/build-qt5-env
-THREADS=-j30
-NEW_QTDIR=/usr/local/Trolltech/Qt5/Qt-5.0.0-$QT_WEEKLY_REV
 
-rm -rf $NEW_QTDIR
+THREADS=
+if [ $N_THREADS -gt 1 ]; then
+    THREADS=-j$N_THREADS
+fi
+
+NEW_QTDIR=
+INSTALL_TYPE=
+if [ $DEVELOPER_BUILD ]; then
+    NEW_QTDIR=$ABSDIR/qt5/qtbase
+    INSTALL_TYPE=-developer-build
+else
+    NEW_QTDIR=/usr/local/Trolltech/Qt5/Qt-5.0.0-$QT_WEEKLY_REV
+    INSTALL_TYPE="-prefix $NEW_QTDIR"
+    rm -rf $NEW_QTDIR
+fi
 
 if [ ! -d qt5 ]
 then
@@ -33,9 +77,9 @@ echo ==========================================================
 export QTDIR=$NEW_QTDIR
 export PATH=$QTDIR/bin:$PATH
 
-./configure -opensource -confirm-license -no-pch -nomake examples -nomake demos -nomake tests -no-gtkstyle -nomake translations -qt-zlib -qt-libpng -qt-libjpeg -qt-sql-sqlite -release -prefix $QTDIR
+./configure -opensource -confirm-license -no-pch -nomake examples -nomake demos -nomake tests -no-gtkstyle -nomake translations -qt-zlib -qt-libpng -qt-libjpeg -qt-sql-sqlite $BUILD_TYPE $INSTALL_TYPE
 
-cd qtbase && make $THREADS && make install && cd ..
+cd qtbase && make $THREADS && if [ ! $DEVELOPER_BUILD ]; then make install; fi && cd ..
 if [ $? -ne 0 ] ; then
   echo FAIL: building qtbase
   exit 1
@@ -43,7 +87,7 @@ fi
 
 for module in $QT5_MODULES
 do
-  cd $module && qmake && make $THREADS && make install && cd ..
+  cd $module && qmake && make $THREADS && if [ ! $DEVELOPER_BUILD ]; then make install; fi && cd ..
   if [ $? -ne 0 ] ; then
     echo FAIL: building $module.
     exit 1
